feat(rds-signer): add RDS Signer (#1823)
