--- lib/lib-rds/README.md
@@ -1,6 +1,0 @@
-# @aws-sdk/lib-rds
-
-[![NPM version](https://img.shields.io/npm/v/@aws-sdk/lib-rds/latest.svg)](https://www.npmjs.com/package/@aws-sdk/lib-rds)
-[![NPM downloads](https://img.shields.io/npm/dm/@aws-sdk/lib-rds.svg)](https://www.npmjs.com/package/@aws-sdk/lib-rds)
-
-## Overview
