--- lib/lib-rds/package.json
@@ -1,55 +1,0 @@
-{
-  "name": "@aws-sdk/lib-rds",
-  "version": "3.43.0",
-  "description": "RDS utility for generating a password that can be used for IAM authentication to an RDS DB.",
-  "main": "./dist-cjs/index.js",
-  "module": "./dist-es/index.js",
-  "types": "./dist-types/index.d.ts",
-  "scripts": {
-    "build": "yarn build:cjs && yarn build:es && yarn build:types",
-    "build:cjs": "tsc -p tsconfig.cjs.json",
-    "build:es": "tsc -p tsconfig.es.json",
-    "build:types": "tsc -p tsconfig.types.json",
-    "downlevel-dts": "downlevel-dts dist-types dist-types/ts3.4",
-    "test": "jest"
-  },
-  "engines": {
-    "node": ">= 10.0.0"
-  },
-  "author": {
-    "name": "AWS SDK for JavaScript Team",
-    "url": "https://aws.amazon.com/javascript/"
-  },
-  "license": "Apache-2.0",
-  "dependencies": {
-    "@aws-sdk/signature-v4": "3.40.0",
-    "@aws-sdk/protocol-http": "3.40.0",
-    "tslib": "^2.3.0"
-  },
-  "peerDependencies": {
-    "@aws-sdk/types": "^3.0.0"
-  },
-  "devDependencies": {
-    "@aws-sdk/types": "3.40.0",
-    "@types/jest": "^26.0.4",
-    "@types/node": "^14.11.2",
-    "jest": "^26.4.2",
-    "typescript": "~4.3.5"
-  },
-  "typesVersions": {
-    "<4.0": {
-      "dist-types/*": [
-        "dist-types/ts3.4/*"
-      ]
-    }
-  },
-  "files": [
-    "dist-*"
-  ],
-  "homepage": "https://github.com/aws/aws-sdk-js-v3/tree/main/lib/lib-rds",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/aws/aws-sdk-js-v3.git",
-    "directory": "lib/lib-rds"
-  }
-}
