--- lib/lib-rds/src/index.spec.ts
@@ -1,7 +1,0 @@
-import * as libRds from "./index";
-
-describe("lib-rds", () => {
-  it("has Signer", () => {
-    expect(libRds.Signer).toBeDefined();
-  });
-});
