+++ packages/rds-signer/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@aws-sdk/util-rds",
+  "name": "@aws-sdk/rds-signer",
   "version": "3.43.0",
   "description": "RDS utility for generating a password that can be used for IAM authentication to an RDS DB.",
   "main": "./dist-cjs/index.js",
@@ -22,10 +22,9 @@
   },
   "license": "Apache-2.0",
   "dependencies": {
-    "@aws-crypto/sha256-browser": "^2.0.0",
-    "@aws-crypto/sha256-js": "^2.0.0",
-    "@aws-sdk/credential-provider-node": "3.40.0",
-    "@aws-sdk/hash-node": "^3.40.0",
+    "@aws-crypto/sha256-browser": "2.0.0",
+    "@aws-crypto/sha256-js": "2.0.0",
+    "@aws-sdk/hash-node": "3.40.0",
     "@aws-sdk/protocol-http": "3.40.0",
     "@aws-sdk/signature-v4": "3.40.0",
     "@aws-sdk/util-format-url": "3.40.0",
@@ -57,10 +56,10 @@
   "react-native": {
     "./dist-es/runtimeConfig": "./dist-es/runtimeConfig.native"
   },
-  "homepage": "https://github.com/aws/aws-sdk-js-v3/tree/main/packages/util-rds",
+  "homepage": "https://github.com/aws/aws-sdk-js-v3/tree/main/packages/rds-signer",
   "repository": {
     "type": "git",
     "url": "https://github.com/aws/aws-sdk-js-v3.git",
-    "directory": "packages/util-rds"
+    "directory": "packages/rds-signer"
   }
 }
