+++ packages/rds-signer/src/Signer.ts
