+++ packages/rds-signer/src/index.spec.ts
@@ -1,0 +1,7 @@
+import * as rdsSigner from "./index";
+
+describe("rds-signer", () => {
+  it("has Signer", () => {
+    expect(rdsSigner.Signer).toBeDefined();
+  });
+});
