+++ packages/rds-signer/src/runtimeConfig.browser.ts
