+++ packages/rds-signer/tsconfig.es.json
