--- packages/util-rds/jest.config.js
@@ -1,5 +1,0 @@
-const base = require("../../jest.config.base.js");
-
-module.exports = {
-  ...base,
-};
