--- packages/util-rds/src/index.spec.ts
@@ -1,7 +1,0 @@
-import * as libRds from "./index";
-
-describe("util-rds", () => {
-  it("has Signer", () => {
-    expect(libRds.Signer).toBeDefined();
-  });
-});
