--- packages/util-rds/tsconfig.cjs.json
@@ -1,9 +1,0 @@
-{
-  "compilerOptions": {
-    "baseUrl": ".",
-    "outDir": "dist-cjs",
-    "rootDir": "src"
-  },
-  "extends": "../../tsconfig.cjs.json",
-  "include": ["src/"]
-}
