--- packages/util-rds/tsconfig.es.json
@@ -1,10 +1,0 @@
-{
-  "compilerOptions": {
-    "baseUrl": ".",
-    "lib": ["es5", "es2015.collection", "dom"],
-    "outDir": "dist-es",
-    "rootDir": "src"
-  },
-  "extends": "../../tsconfig.es.json",
-  "include": ["src/"]
-}
