--- packages/util-rds/tsconfig.types.json
@@ -1,9 +1,0 @@
-{
-  "compilerOptions": {
-    "baseUrl": ".",
-    "declarationDir": "dist-types",
-    "rootDir": "src"
-  },
-  "extends": "../../tsconfig.types.json",
-  "include": ["src/"]
-}
